Cast index ints to const int to help msvc pick a method for the overloaded [] operator (fixes errors on 32b MSVC8)
svn:revision:51718
svn:branch:trunk
svn:account:erikgreenwald

+++ src/librt/primitives/dsp/dsp_brep.cpp
@@ -408,14 +408,14 @@ rt_dsp_brep(ON_Brep **b, const struct rt_db_internal *ip, const struct bn_tol *)
     bool exceedtol = false;
     ON_SimpleArray<ON_3dPointArray *> bezcurvarray(dsp_ip->dsp_ycnt);
     for (unsigned int y = 0; y < dsp_ip->dsp_ycnt; y++) {
-	bezcurvarray[y] = new ON_3dPointArray(dsp_ip->dsp_xcnt);
+	bezcurvarray[(const int)y] = new ON_3dPointArray(dsp_ip->dsp_xcnt);
 	for (unsigned int x = 0; x < dsp_ip->dsp_xcnt; x++) {
-	    bezcurvarray[y]->Append(bezsurf->CV(x, y));
+	    bezcurvarray[(const int)y]->Append(bezsurf->CV(x, y));
 	}
     }
     while (currentdegree && !exceedtol) {
 	for (unsigned int y = 0; y < dsp_ip->dsp_ycnt; y++) {
-	    if (DegreeReduction(currentdegree - 1, *bezcurvarray[y], tol, maxerr) == -1) {
+	    if (DegreeReduction(currentdegree - 1, *bezcurvarray[(const int)y], tol, maxerr) == -1) {
 		exceedtol = true;
 		bu_log("%lf\n", maxerr);
 		break;
@@ -425,7 +425,7 @@ rt_dsp_brep(ON_Brep **b, const struct rt_db_internal *ip, const struct bn_tol *)
 	    currentdegree--;
 	    for (unsigned int y = 0; y < dsp_ip->dsp_ycnt; y++) {
 		for (unsigned int x = 0; x <= currentdegree; x++) {
-		    reducedsurf->SetCV(x, y, (*bezcurvarray[y])[x]);
+		    reducedsurf->SetCV(x, y, (*bezcurvarray[y])[(const int)x]);
 		}
 	    }
 	}
@@ -459,7 +459,7 @@ rt_dsp_brep(ON_Brep **b, const struct rt_db_internal *ip, const struct bn_tol *)
 	    currentdegree--;
 	    for (int x = 0; x < reducedsurf->m_order[0]; x++) {
 		for (unsigned int y = 0; y <= currentdegree; y++) {
-		    reducedsurf->SetCV(x, y, (*bezcurvarray[x])[y]);
+		    reducedsurf->SetCV(x, y, (*bezcurvarray[(const int)x])[(const int)y]);
 		}
 	    }
 	}
